warehouse/requirements.txt
six
pyyaml
requests
Django>=2.0.0
djangorestframework>=3.0,<4.0

